parser: allow for default generator
