+++ src/parse.rs
@@ -1,5 +1,5 @@
 // Call `apply` macro on this generated form of token tree;
-// $ctx, { day DAY { gen GENERATOR { { sol SOLUTION } { sol SOLUTION } } } }
+// $ctx, { day DAY { { gen GENERATOR } { { sol SOLUTION } { sol SOLUTION } } } }
 #[macro_export]
 macro_rules! parse {
     // Read day
@@ -7,7 +7,15 @@ macro_rules! parse {
         day $apply: ident, $ctx: tt, $val: expr;
         $day: ident : $generator: ident => $( $tail: tt )*
     ) => {
-        $crate::parse!( sol $apply, $ctx, $val; { day $day { gen $generator { } } }; $( $tail )* )
+        $crate::parse!( sol $apply, $ctx, $val; { day $day { { gen $generator } { } } }; $( $tail )* )
+    };
+
+    // Read day: default generator
+    (
+        day $apply: ident, $ctx: tt, $val: expr;
+        $day: ident => $( $tail: tt )*
+    ) => {
+        $crate::parse!( sol $apply, $ctx, $val; { day $day { { gen_default } { } } }; $( $tail )* )
     };
 
     // Empty rules
@@ -16,12 +24,12 @@ macro_rules! parse {
     // Read solution
     (
         sol $apply: ident, $ctx: tt, $val: expr;
-        { day $day: tt { gen $generator: tt { $( $acc: tt )* } } } ;
+        { day $day: tt { $gen: tt { $( $acc: tt )* } } } ;
         $sol: ident $( $tail: tt )*
     ) => {
         $crate::parse!(
             post_sol $apply, $ctx, $val;
-            { day $day { gen $generator { $( $acc )* { sol $sol } } } };
+            { day $day { $gen { $( $acc )* { sol $sol } } } };
             $( $tail )*
         )
     };
